buffer to geometry: refactor
