+++ src/plugins/score-plugin-threedim/Threedim/BufferToGeometry.cpp
@@ -12,87 +12,6 @@ namespace Threedim
 namespace
 {
 
-[[nodiscard]] constexpr int32_t
-attributeFormatSize(BuffersToGeometry::AttributeFormat fmt) noexcept
-{
-  return Threedim::attributeFormatSize((halp::attribute_format)fmt);
-}
-
-[[nodiscard]] constexpr halp::attribute_format
-toHalpFormat(BuffersToGeometry::AttributeFormat fmt) noexcept
-{
-  return static_cast<halp::attribute_format>(fmt);
-}
-
-[[nodiscard]] constexpr halp::primitive_topology
-toHalpTopology(BuffersToGeometry::PrimitiveTopology t) noexcept
-{
-  switch(t)
-  {
-    case BuffersToGeometry::Triangles:
-      return halp::primitive_topology::triangles;
-    case BuffersToGeometry::TriangleStrip:
-      return halp::primitive_topology::triangle_strip;
-    case BuffersToGeometry::TriangleFan:
-      return halp::primitive_topology::triangle_fan;
-    case BuffersToGeometry::Lines:
-      return halp::primitive_topology::lines;
-    case BuffersToGeometry::LineStrip:
-      return halp::primitive_topology::line_strip;
-    case BuffersToGeometry::Points:
-      return halp::primitive_topology::points;
-  }
-  return halp::primitive_topology::triangles;
-}
-
-[[nodiscard]] constexpr halp::cull_mode
-toHalpCullMode(BuffersToGeometry::CullMode c) noexcept
-{
-  switch(c)
-  {
-    case BuffersToGeometry::None:
-      return halp::cull_mode::none;
-    case BuffersToGeometry::Front:
-      return halp::cull_mode::front;
-    case BuffersToGeometry::Back:
-      return halp::cull_mode::back;
-  }
-  return halp::cull_mode::none;
-}
-
-[[nodiscard]] constexpr halp::front_face
-toHalpFrontFace(BuffersToGeometry::FrontFace f) noexcept
-{
-  switch(f)
-  {
-    case BuffersToGeometry::CounterClockwise:
-      return halp::front_face::counter_clockwise;
-    case BuffersToGeometry::Clockwise:
-      return halp::front_face::clockwise;
-  }
-  return halp::front_face::counter_clockwise;
-}
-
-[[nodiscard]] constexpr halp::index_format
-toHalpIndexFormat(BuffersToGeometry::IndexFormat f) noexcept
-{
-  switch(f)
-  {
-    case BuffersToGeometry::UInt16:
-      return halp::index_format::uint16;
-    case BuffersToGeometry::UInt32:
-      return halp::index_format::uint32;
-  }
-  return halp::index_format::uint32;
-}
-
-[[nodiscard]] constexpr halp::binding_classification
-toHalpClassification(bool instanced) noexcept
-{
-  return instanced ? halp::binding_classification::per_instance
-                   : halp::binding_classification::per_vertex;
-}
-
 // Convert location int to attribute_semantic enum
 [[nodiscard]] constexpr halp::attribute_semantic
 toHalpLocation(int32_t loc, bool instanced) noexcept
