+++ src/plugins/score-plugin-threedim/Threedim/BufferToGeometry2.cpp
@@ -11,103 +11,6 @@
 namespace Threedim
 {
 
-namespace
-{
-
-[[nodiscard]] constexpr int32_t
-attributeFormatSize(BuffersToGeometry2::AttributeFormat fmt) noexcept
-{
-  return Threedim::attributeFormatSize((halp::attribute_format)fmt);
-}
-
-[[nodiscard]] constexpr halp::attribute_format
-toHalpFormat(BuffersToGeometry2::AttributeFormat fmt) noexcept
-{
-  return static_cast<halp::attribute_format>(fmt);
-}
-
-[[nodiscard]] constexpr halp::primitive_topology
-toHalpTopology(BuffersToGeometry2::PrimitiveTopology t) noexcept
-{
-  switch(t)
-  {
-    case BuffersToGeometry2::Triangles:
-      return halp::primitive_topology::triangles;
-    case BuffersToGeometry2::TriangleStrip:
-      return halp::primitive_topology::triangle_strip;
-    case BuffersToGeometry2::TriangleFan:
-      return halp::primitive_topology::triangle_fan;
-    case BuffersToGeometry2::Lines:
-      return halp::primitive_topology::lines;
-    case BuffersToGeometry2::LineStrip:
-      return halp::primitive_topology::line_strip;
-    case BuffersToGeometry2::Points:
-      return halp::primitive_topology::points;
-  }
-  return halp::primitive_topology::triangles;
-}
-
-[[nodiscard]] constexpr halp::cull_mode
-toHalpCullMode(BuffersToGeometry2::CullMode c) noexcept
-{
-  switch(c)
-  {
-    case BuffersToGeometry2::None:
-      return halp::cull_mode::none;
-    case BuffersToGeometry2::Front:
-      return halp::cull_mode::front;
-    case BuffersToGeometry2::Back:
-      return halp::cull_mode::back;
-  }
-  return halp::cull_mode::none;
-}
-
-[[nodiscard]] constexpr halp::front_face
-toHalpFrontFace(BuffersToGeometry2::FrontFace f) noexcept
-{
-  switch(f)
-  {
-    case BuffersToGeometry2::CounterClockwise:
-      return halp::front_face::counter_clockwise;
-    case BuffersToGeometry2::Clockwise:
-      return halp::front_face::clockwise;
-  }
-  return halp::front_face::counter_clockwise;
-}
-
-[[nodiscard]] constexpr halp::index_format
-toHalpIndexFormat(BuffersToGeometry2::IndexFormat f) noexcept
-{
-  switch(f)
-  {
-    case BuffersToGeometry2::UInt16:
-      return halp::index_format::uint16;
-    case BuffersToGeometry2::UInt32:
-      return halp::index_format::uint32;
-  }
-  return halp::index_format::uint32;
-}
-
-[[nodiscard]] constexpr halp::binding_classification
-toHalpClassification(bool instanced) noexcept
-{
-  return instanced ? halp::binding_classification::per_instance
-                   : halp::binding_classification::per_vertex;
-}
-
-// Resolve a user-provided semantic name to ossia::attribute_semantic.
-// Uses case-insensitive name_to_semantic; if unrecognized, returns custom.
-[[nodiscard]] int resolveSemanticFromName(const std::string& name) noexcept
-{
-  if(name.empty())
-    return static_cast<int>(ossia::attribute_semantic::custom);
-
-  auto sem = ossia::name_to_semantic(name);
-  return static_cast<int>(sem);
-}
-
-} // anonymous namespace
-
 BuffersToGeometry2::BuffersToGeometry2()
 {
   // Initialize transform to identity
